DEV : removing old _sample module to sampling
