--- src/relife/stochastic_processes/_sample/__init__.py
@@ -1,13 +1,0 @@
-from ._iterables import (
-    Kijima1ProcessIterable,
-    Kijima2ProcessIterable,
-    NonHomogeneousPoissonProcessIterable,
-    RenewalProcessIterable,
-)
-
-__all__ = [
-    "NonHomogeneousPoissonProcessIterable",
-    "RenewalProcessIterable",
-    "Kijima1ProcessIterable",
-    "Kijima2ProcessIterable",
-]
